Allow to set the initial admin password
Add initialAdminPasswordFile as an option to specify where to read the initial
admin password from. This allows to use the deployment keys machinery to inject
the initial admin password in a safe way.

It does fallback to a random password which is stored in root's home if no key
is set.