Add patch for enterprise to allow to sneak in environment variables
Faced the issue of how to provide a password e.g. for the database connection
without ending up putting this into the nix store. The problem with the store
would be that things are world-readable there.

The approach is inspired by Plaster from the Pylons project. I opted to use the
format string syntax since ConfigParser does have trouble if I would use the
string interpolation syntax.