Move opt symlink logic into the nixos modules
Turns out that this is the correct place for the implementation. Reason is that
a user can set the option "package" and the symlink has to be based on this
value. This knowledge is not available within the release file.

Using environment.systemPackages and environment.pathsToLink in order to get the
desired results.

The configuration file should not need any linking, since it is available under
"/run/keys/...".